requirements.txt
# Core dependencies
flask==2.0.1
flask-cors==3.0.10
pandas==1.3.3
numpy==1.21.2
tqdm==4.62.3
python-dotenv==0.19.1
joblib==1.1.0
scikit-learn==1.0

# Reddit scraping
praw==7.5.0

# NLP and ML
spacy==3.1.3
torch==1.10.0
transformers==4.11.3
sentence-transformers==2.1.0

# Download spaCy models with:
# python -m spacy download en_core_web_md